// src/api/types/NonCompulsoryHealthInsuranceResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type NonCompulsoryHealthInsuranceResponse = {
  /**
   * @type integer | undefined, int64
   */
  id?: number;
  /**
   * @type integer | undefined, int64
   */
  userId?: number;
  /**
   * @type string | undefined
   */
  insuranceNumber?: string;
  /**
   * @type string | undefined
   */
  placeOfRegistration?: string;
  /**
   * @type string | undefined, date
   */
  expiryDate?: string;
  /**
   * @type string | undefined, date-time
   */
  createdAt?: string;
  /**
   * @type string | undefined, date-time
   */
  updatedAt?: string;
};